New script to setup a local repo for offline updates
With this script, one can setup a local repository from an archive which
contains the files of a repository, and use it to install updates or
additional packages.

The same script is meant to be used to update the repository. In order
not to use too much space on the disk, we accept that the operation
makes the repository unusable for the duration of the local copy.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>