Build protected targets from the latest commit of their intended branch
Ensures that packages for protected targets (e.g., v8.3-incoming from
master, v8.2-incoming from 8.2) are built directly from the latest
commit of their respective target branch.

This prevent from building a package from sources that may not be in
the intended branch.

This prevents another common issue where a PR is merged into the target
branch but the build uses the PR branch's commit. Such builds omit the
merge commit, leading to the PR appearing in subsequent build reports
despite being effectively integrated.

A --force option is added to bypass these checks, for the few cases
for which it is valid to not build the last commit of the intended
branch.

Signed-off-by: Gaëtan Lehmann <gaetan.lehmann@vates.tech>