import_srpm: check command rpm2cpio|cpio
`rpm2cpio` might not exist on the system, and `cpio` return the error
"premature end of archive" in that case.

This doesn't check if `rpm2cpio` returns an error but it's probably ok
to lean on `cpio` returning an error.

Also, check early if `rpm2cpio` actually exist on the system.

Signed-off-by: Anthony PERARD <anthony.perard@vates.tech>