koji_build: Add --xapi-build subcommand
Like --pre-build or --test-build, it takes an additional argument. Here
it points to a commit reference in the xcp-ng/xen-api repository:

koji_build.py --xapi-build 1b20f51 v8.3-u-asultanov2 .

The script will create an appropriate tar.gz archive, replacing the
current Source0 with it, and increment the release tag accordingly.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>