Add a script to terminate idle pg connections
Koji keeps opening new connections to the database, and keeps them open
up to the point where no new connection can be open. Koji then starts
to fail, with a database outage error message.

This scripts terminates the connexions older that 10 minutes, and is
intended to run every minute on the koji server.

Signed-off-by: Gaëtan Lehmann <gaetan.lehmann@vates.tech>