koji: Add supported git branches (eg: [8.3, master] for v8.3-* tags)
make `koji_build` accept builds for v8.3 and from both `master` and `8.3`

Preparation for switch `master` to be "next" aka v9

To my understanding, the motivation of change is part of a bigger
plan, to prepare a smooth transition to enable bleeding edge
development on master branch. Of course maintenance will remain in
v8.x branches.

Once the master branch needs to be stabilized then a new 9.0 branch
can be open (this mean declared in PROTECTED_TARGETS, and also move
the `master` branch from `8.3` target to `9.0` for a future cycle)

Meanwhile the development on master is still open and might become the
base for the next cycle (TBD 9.1 or 10).

Note that `8.3` git branches are not created yet in repos, it would
make sense to open them soon or later, ideally when master branches
need to receive feature from next release. Coincidentaly thoses 8.3
branches will be used for maintenance of the LTS release and then
developers/RE might consider back-porting fixes from master to 8.3.

What is unclear to me in the plan is when the new koji target/tag v9
will be defined, or if an `unstable` target will be added before,
probably this will have to be coordinated with CI.

Signed-off-by: Philippe Coval <philippe.coval@vates.tech>